+++ common/src/main/java/com/faboslav/structurify/common/config/client/api/option/OptionPair.java
@@ -1,7 +1,3 @@
 package com.faboslav.structurify.common.config.client.api.option;
 
-import dev.isxander.yacl3.api.Option;
-
-public record OptionPair<K extends Option<?>, V extends Option<?>>(K firstOption, V secondOption)
-{
-}
+public record OptionPair<K, V>(K firstOption, V secondOption) {}